--- cmd/constants.go
@@ -1,9 +1,0 @@
-// Copyright (C) 2022, Ava Labs, Inc. All rights reserved.
-// See the file LICENSE for licensing terms.
-package cmd
-
-const (
-	maxLogFileSize   = 4
-	maxNumOfLogFiles = 5
-	retainOldFiles   = 0 // retain all old log files
-)
