+++ internal/mocks/bin_checker.go
