+++ internal/mocks/bin_downloader.go
