+++ internal/mocks/proc_checker.go
